Refactor: Rename CommandRunner to CLICommandHandler. Rename CommandRequest to CLIRequest.
