+++ 01-shapes/src/main/java/com/pedromg/bluej/shapes/command/Command.java
@@ -7,6 +7,6 @@ public interface Command {
    *
    * @param request command line request
    */
-  void execute(CommandRequest request);
+  void execute(CLIRequest request);
 
 }
